requirements.txt
python-dotenv>=1.0.1
openai>=1.50.2
pandas>=2.2.2
sqlglot>=25.6.0
sqlite-utils>=3.36
tabulate>=0.9.0
matplotlib>=3.8.4
streamlit>=1.37.0
